Bug 1098967 - Use native Win32 APIs to append to the LogAlloc output. r=nfroyd
Because it turns out that the POSIX API the CRT exposes doesn't do O_APPEND
in a sane manner.

UltraBlame original commit: 4405339504c9d8f0dfc3234901916dd7895a0b96

+++ memory/replace/logalloc/FdPrintf.h
@@ -13,7 +13,11 @@
 
 
 
-extern void FdPrintf(int aFd, const char* aFormat, ...)
+
+
+
+
+extern void FdPrintf(intptr_t aFd, const char* aFormat, ...)
 #ifdef __GNUC__
 __attribute__((format(printf, 2, 3)))
 #endif
